requirements.txt
pytest~=8.3.5
selenium~=3.141.0
python-dotenv~=1.0.1
playwright~=1.52.0
requests~=2.32.3
pip~=21.2.4
configparser~=5.0.2
attrs~=23.2.0
wheel~=0.37.0
pyee~=13.0.0
exceptiongroup~=1.2.2
idna~=3.7
PyYAML~=6.0.2
Naked~=0.1.32
crypto~=1.4.1
pluggy~=1.5.0
iniconfig~=2.1.0
certifi~=2024.7.4
urllib3~=1.26.4
greenlet~=3.2.2
charset-normalizer~=3.3.2
simplejson~=3.19.2
packaging~=24.1